+++ mayan/apps/documents/tests/mixins/document_mixins.py
@@ -9,9 +9,9 @@ from ...literals import DOCUMENT_FILE_ACTION_PAGES_NEW, PAGE_RANGE_ALL
 from ...models import Document, DocumentType
 
 from ..literals import (
-    DEFAULT_DOCUMENT_STUB_LABEL, TEST_DOCUMENT_DESCRIPTION_EDITED,
-    TEST_DOCUMENT_TYPE_LABEL, TEST_SMALL_DOCUMENT_FILENAME,
-    TEST_SMALL_DOCUMENT_PATH
+    DEFAULT_DOCUMENT_STUB_LABEL, TEST_DOCUMENT_DESCRIPTION,
+    TEST_DOCUMENT_DESCRIPTION_EDITED, TEST_DOCUMENT_TYPE_LABEL,
+    TEST_SMALL_DOCUMENT_FILENAME, TEST_SMALL_DOCUMENT_PATH
 )
 
 
@@ -147,16 +147,25 @@ class DocumentTestMixin:
                 'sample_documents', self.test_document_file_filename
             )
 
-    def _upload_test_document(self, document_type=None, label=None, _user=None):
+    def _upload_test_document(
+        self, description=None, document_file_attributes=None,
+        document_type=None, document_version_attributes=None, label=None,
+        _user=None
+    ):
         self._calculate_test_document_path()
 
         if not label:
             label = self.test_document_filename
 
+        test_document_description = description or '{}_{}'.format(
+            TEST_DOCUMENT_DESCRIPTION, len(self.test_documents)
+        )
+
         document_type = document_type or self.test_document_type
 
         with open(file=self.test_document_path, mode='rb') as file_object:
             document, document_file = document_type.new_document(
+                description=test_document_description,
                 file_object=file_object, label=label,
                 language=self.test_document_language, _user=_user
             )
@@ -168,6 +177,18 @@ class DocumentTestMixin:
         self.test_document_file_page = document_file.file_pages.first()
         self.test_document_version = self.test_document.version_active
         self.test_document_version_page = self.test_document_version.version_pages.first()
+
+        if document_file_attributes:
+            for key, value in document_file_attributes.items():
+                setattr(self.test_document_file, key, value)
+
+            self.test_document_file.save()
+
+        if document_version_attributes:
+            for key, value in document_version_attributes.items():
+                setattr(self.test_document_version, key, value)
+
+            self.test_document_version.save()
 
     def _upload_test_document_file(self, action=None, _user=None):
         self._calculate_test_document_file_path()