+++ mayan/apps/dynamic_search/classes.py
@@ -11,16 +11,14 @@ from django.utils.module_loading import import_string
 from django.utils.translation import ugettext as _
 
 from mayan.apps.common.class_mixins import AppsModuleLoaderMixin
-from mayan.apps.common.exceptions import ResolverPipelineError
-from mayan.apps.common.utils import (
-    ResolverPipelineModelAttribute, flatten_list, get_class_full_name
-)
+from mayan.apps.common.utils import get_class_full_name
 from mayan.apps.views.literals import LIST_MODE_CHOICE_LIST
 
 from .exceptions import DynamicSearchException
 from .literals import (
-    DEFAULT_SCOPE_ID, DELIMITER, SCOPE_MATCH_ALL, SCOPE_MARKER,
-    SCOPE_OPERATOR_CHOICES, SCOPE_OPERATOR_MARKER, SCOPE_RESULT_MAKER
+    DEFAULT_SCOPE_ID, DELIMITER, MESSAGE_FEATURE_NO_STATUS, SCOPE_MATCH_ALL,
+    SCOPE_MARKER, SCOPE_OPERATOR_CHOICES, SCOPE_OPERATOR_MARKER,
+    SCOPE_RESULT_MAKER
 )
 from .settings import (
     setting_backend, setting_backend_arguments,
@@ -30,18 +28,54 @@ logger = logging.getLogger(name=__name__)
 
 
 class SearchBackend:
-    @staticmethod
-    def get_instance(extra_kwargs=None):
-        kwargs = setting_backend_arguments.value.copy()
-        if extra_kwargs:
-            kwargs.update(extra_kwargs)
-
-        return import_string(dotted_path=setting_backend.value)(
-            **kwargs
-        )
+    _initialized = False
+    _search_field_transformations = {}
 
     @staticmethod
-    def initialize():
+    def _disable():
+        for search_model in SearchModel.all():
+            post_save.disconnect(
+                dispatch_uid='search_handler_index_instance',
+                sender=search_model.model
+            )
+            pre_delete.disconnect(
+                dispatch_uid='search_handler_deindex_instance',
+                sender=search_model.model
+            )
+
+            for proxy in search_model.proxies:
+                post_save.disconnect(
+                    dispatch_uid='search_handler_index_instance',
+                    sender=proxy
+                )
+                pre_delete.disconnect(
+                    dispatch_uid='search_handler_deindex_instance',
+                    sender=proxy
+                )
+
+            for related_model, path in search_model.get_related_models():
+                post_save.disconnect(
+                    dispatch_uid='search_handler_index_related_instance_{}_{}'.format(
+                        get_class_full_name(klass=search_model.model),
+                        get_class_full_name(klass=related_model)
+                    ), sender=related_model
+                )
+                pre_delete.disconnect(
+                    dispatch_uid='search_handler_index_related_instance_delete_{}_{}'.format(
+                        get_class_full_name(klass=search_model.model),
+                        get_class_full_name(klass=related_model)
+                    ), sender=related_model
+                )
+
+        for through_model, data in SearchModel.get_through_models().items():
+            m2m_changed.disconnect(
+                dispatch_uid='search_handler_index_related_instance_m2m_{}'.format(
+                    get_class_full_name(klass=through_model),
+                ), sender=through_model
+            )
+
+    @staticmethod
+    def _enable():
         # Hidden import.
         from .handlers import (
             handler_deindex_instance, handler_index_instance,
@@ -104,57 +138,24 @@ class SearchBackend:
             )
 
     @staticmethod
+    def get_instance(extra_kwargs=None):
+        kwargs = setting_backend_arguments.value.copy()
+        if extra_kwargs:
+            kwargs.update(extra_kwargs)
+
+        return import_string(dotted_path=setting_backend.value)(
+            **kwargs
+        )
+
+    @staticmethod
     def limit_queryset(queryset):
         pk_list = queryset.values('pk')[:setting_results_limit.value]
         return queryset.filter(pk__in=pk_list)
 
-    @staticmethod
-    def terminate():
-        for search_model in SearchModel.all():
-            post_save.disconnect(
-                dispatch_uid='search_handler_index_instance',
-                sender=search_model.model
-            )
-            pre_delete.disconnect(
-                dispatch_uid='search_handler_deindex_instance',
-                sender=search_model.model
-            )
-
-            for proxy in search_model.proxies:
-                post_save.disconnect(
-                    dispatch_uid='search_handler_index_instance',
-                    sender=proxy
-                )
-                pre_delete.disconnect(
-                    dispatch_uid='search_handler_deindex_instance',
-                    sender=proxy
-                )
-
-            for related_model, path in search_model.get_related_models():
-                post_save.disconnect(
-                    dispatch_uid='search_handler_index_related_instance_{}_{}'.format(
-                        get_class_full_name(klass=search_model.model),
-                        get_class_full_name(klass=related_model)
-                    ), sender=related_model
-                )
-                pre_delete.disconnect(
-                    dispatch_uid='search_handler_index_related_instance_delete_{}_{}'.format(
-                        get_class_full_name(klass=search_model.model),
-                        get_class_full_name(klass=related_model)
-                    ), sender=related_model
-                )
-
-        for through_model, data in SearchModel.get_through_models().items():
-            m2m_changed.disconnect(
-                dispatch_uid='search_handler_index_related_instance_m2m_{}'.format(
-                    get_class_full_name(klass=through_model),
-                ), sender=through_model
-            )
-
     def __init__(self, **kwargs):
         self.kwargs = kwargs
 
-    def _search(self, global_and_search, query, search_model, user):
+    def _search(self, global_and_search, query, search_model, user, ignore_limit):
         raise NotImplementedError
 
     def cleanup_query(self, query, search_model):
@@ -176,10 +177,14 @@ class SearchBackend:
 
         return clean_query
 
+    def close(self):
+        """
+        Optional method to terminate a backend instance, such as closing
+        connections.
+        """
+
     def decode_query(self, query, global_and_search=False):
         # Clean up the query.
-        # The original query is immutable, create a new
-        # mutable copy.
         query.pop('_match_all', None)
 
         # Turn scoped query dictionary into a series of unscoped queries.
@@ -217,7 +222,7 @@ class SearchBackend:
                     if key.endswith(SCOPE_MATCH_ALL):
                         scope_id, key = key.split(DELIMITER, 1)
                         scopes.setdefault(scope_id, {})
-                        scope_match_all = value.upper() == 'TRUE'
+                        scope_match_all = value.lower() in ['on', 'true']
                         scopes[scope_id]['match_all'] = scope_match_all
                     else:
                         # Must be a scoped query.
@@ -230,7 +235,7 @@ class SearchBackend:
                         scopes[scope_id]['query'][key] = value
             else:
                 scopes.setdefault(scope_id, {})
-                scopes[scope_id].setdefault('match_all', scope_match_all)
+                scopes[scope_id].setdefault('match_all', global_and_search)
                 scopes[scope_id].setdefault('query', {})
                 scopes[scope_id]['query'][key] = value
         else:
@@ -245,10 +250,83 @@ class SearchBackend:
         }
 
     def deindex_instance(self, instance):
-        raise NotImplementedError
+        """
+        Optional method to remove an model instance from the search index.
+        """
+
+    def get_resolved_field_map(self, search_model):
+        result = {}
+        for search_field in self.get_search_model_fields(search_model=search_model):
+            backend_field_type = self.field_map.get(
+                search_field.field_type
+            )
+
+            if backend_field_type:
+                result[search_field.get_full_name()] = backend_field_type
+            else:
+                logger.warning(
+                    'Unknown field type "%s" for model "%s"',
+                    search_field.get_full_name(),
+                    search_model.get_full_name()
+                )
+
+        return result
+
+    def get_search_field_transformation(self, search_field):
+        if search_field not in self.__class__._search_field_transformations:
+            field_map = self.get_resolved_field_map(
+                search_model=search_field.search_model
+            )
+            transformation = field_map[search_field.field].get(
+                'transformation', SearchModel.function_return_same
+            )
+            self.__class__._search_field_transformations[
+                search_field
+            ] = transformation
+
+        return self.__class__._search_field_transformations[search_field]
+
+    def get_search_model_fields(self, search_model):
+        result = search_model.search_fields.copy()
+        result.append(
+            SearchField(search_model=search_model, field='id', label='ID')
+        )
+        return result
+
+    def get_status(self):
+        """
+        Backend specific method to provide status and statistics information.
+        """
+        if not hasattr(self, '_get_status'):
+            return MESSAGE_FEATURE_NO_STATUS
+        else:
+            return self._get_status()
 
     def index_instance(self, instance, exclude_model=None, exclude_kwargs=None):
-        raise NotImplementedError
+        """
+        Optional method to add or update an model instance to the search
+        index.
+        """
+
+    def index_search_model(self, search_model, range_string=None):
+        """
+        Optional method to add or update all instance of a model.
+        """
+
+    def initialize(self):
+        if not self.__class__._initialized:
+            self.__class__._initialized = True
+            self._initialize()
+
+    def _initialize(self):
+        """
+        Optional method to setup the backend. Executed once on every boot up.
+        """
+
+    def reset(self, search_model=None):
+        """
+        Optional method to clear all search indices.
+        """
 
     def search(
         self, query, search_model, user, global_and_search=False
@@ -260,7 +338,6 @@ class SearchBackend:
         result = self.decode_query(
             global_and_search=global_and_search, query=query
         )
-
         # Recursive call to the backend's search using queries as unscoped
         # and then merge then using the corresponding operator.
         queryset = self.solve_scope(
@@ -327,7 +404,18 @@ class SearchBackend:
                         query=query, user=user
                     )
                 else:
-                    return search_model.model._meta.default_manager.none()
+                    return search_model.get_queryset().none()
+
+    def tear_down(self):
+        """
+        Optional method to clean up and/or destroy search backend structures
+        like indices.
+        """
+
+    def upgrade(self):
+        """
+        Optional method to upgrade the search backend persistent structures.
+        """
 
 
 class SearchField:
@@ -342,6 +430,15 @@ class SearchField:
         self.field = field
         self._label = label
         self.transformation_function = transformation_function
+
+    def __repr__(self):
+        return '<{}: {}>'.format(
+            self.__class__.__name__, self.field
+        )
+
+    @cached_property
+    def field_type(self):
+        return self.get_model_field().__class__
 
     def get_full_name(self):
         return self.field
@@ -359,6 +456,14 @@ class SearchField:
     @cached_property
     def model(self):
         return self.search_model.model
+
+    @cached_property
+    def related_model(self):
+        return self.get_model_field().model
+
+    @cached_property
+    def reverse_path(self):
+        return reverse_field_path(model=self.model, path=self.field)[1]
 
 
 class SearchModel(AppsModuleLoaderMixin):
@@ -403,7 +508,7 @@ class SearchModel(AppsModuleLoaderMixin):
     @classmethod
     def get_for_model(cls, instance):
         # Works the same for model classes and model instances.
-        return cls.get(name=instance._meta.label)
+        return cls.get(name=instance._meta.label.lower())
 
     @classmethod
     def get_through_models(cls):
@@ -434,12 +539,15 @@ class SearchModel(AppsModuleLoaderMixin):
         self._label = label
         self.app_label = app_label
         self.list_mode = list_mode or LIST_MODE_CHOICE_LIST
-        self.model_name = model_name
+        self.model_name = model_name.lower()
         self._proxies = []  # Lazy
         self.permission = permission
         self.queryset = queryset
         self.search_fields = []
+        self.search_fields_dict = {}
         self.serializer_path = serializer_path
+
+        self.add_model_field(field='id', label=_('ID'))
 
         self.manager_name = manager_name or self.model._meta.default_manager.name
 
@@ -463,8 +571,10 @@ class SearchModel(AppsModuleLoaderMixin):
         """
         search_field = SearchField(self, *args, **kwargs)
         self.search_fields.append(search_field)
+        self.search_fields_dict[search_field.get_full_name()] = search_field
 
     def add_proxy_model(self, app_label, model_name):
+        model_name = model_name.lower()
         self._proxies.append(
             {
                 'app_label': app_label, 'model_name': model_name
@@ -472,6 +582,30 @@ class SearchModel(AppsModuleLoaderMixin):
         )
 
         self.__class__._registry['{}.{}'.format(app_label, model_name)] = self
+
+    @cached_property
+    def fields_direct(self):
+        result = []
+
+        for search_field in self.search_fields:
+            field_name = search_field.field
+
+            if '__' not in field_name:
+                result.append(search_field)
+
+        return result
+
+    @cached_property
+    def fields_related(self):
+        result = []
+
+        for search_field in self.search_fields:
+            field_name = search_field.field
+
+            if '__' in field_name:
+                result.append(search_field)
+
+        return result
 
     def get_fields_simple_list(self):
         """
@@ -489,7 +623,7 @@ class SearchModel(AppsModuleLoaderMixin):
         return '{}.{}'.format(self.app_label, self.model_name)
 
     def get_queryset(self):
-        if self.queryset:
+        if self.queryset is not None:
             return self.queryset()
         else:
             return self.model._meta.managers_map[self.manager_name].all()
@@ -508,20 +642,19 @@ class SearchModel(AppsModuleLoaderMixin):
 
     def get_search_field(self, full_name):
         try:
-            return self.search_fields[full_name]
+            return self.search_fields_dict[full_name]
         except KeyError:
             raise KeyError('No search field named: %s' % full_name)
-
-    def get_status(self):
-        """
-        Backend specify method to provide status and statistics information.
-        """
 
     @cached_property
     def label(self):
         if not self._label:
             self._label = self.model._meta.verbose_name
         return self._label
+
+    @cached_property
+    def base_model(self):
+        return self.model._meta.proxy_for_model or self.model
 
     @cached_property
     def model(self):
@@ -534,43 +667,55 @@ class SearchModel(AppsModuleLoaderMixin):
         return self.get_full_name()
 
     def populate(
-        self, field_map, instance, exclude_model=None, exclude_kwargs=None
+        self, backend, instance, exclude_model=None, exclude_kwargs=None
     ):
-        """
-        Method that receives an instance and a field map dictionary
-        consisting of attribute names and transformations to apply.
-        Returns a dictionary of the instance values with their respective
-        transformations. Makes it easy to pre process an instance before
-        indexing it.
-        """
-        if exclude_model and exclude_kwargs:
-            resolver_extra_kwargs = {
-                'exclude': exclude_kwargs, 'model': exclude_model
-            }
-        else:
-            resolver_extra_kwargs = {}
-
         result = {}
-        for field in field_map:
-            try:
-                value = ResolverPipelineModelAttribute.resolve(
-                    attribute=field, obj=instance,
-                    resolver_extra_kwargs=resolver_extra_kwargs
+        search_model = SearchModel.get_for_model(instance=instance)
+
+        for direct_field in search_model.fields_direct:
+            field_name = direct_field.field
+
+            # Fetch the fields that produce a finite number of results.
+            value = getattr(instance, field_name)
+
+            result[field_name] = backend.get_search_field_transformation(
+                search_field=direct_field
+            )(value)
+
+        for related_field in search_model.fields_related:
+            # Fetch the fields that produce an undetermined number of
+            # results.
+            # This implementation is a balance between memory usage and
+            # performance.
+            # Attempting to fetch all of these fields at the same time using
+            # joins creates a single query that kills the database and
+            # causes the OOM to kill this method's process due to run away
+            # memory usage.
+            # This implementation is not infinitely scalable but has been
+            # tested up to 500 pages per document version + 500 pages per
+            # document file with an indexing chunk size of 25 for a total
+            # of 25,000 page results per super call.
+            field_name = related_field.field
+            last_field = field_name.split('__')[-1]
+
+            sub_queryset = related_field.related_model._meta.default_manager.filter(
+                **{
+                    '{}'.format(related_field.reverse_path): instance.pk
+                }
+            ).values_list(last_field, flat=True)
+
+            if exclude_model and related_field.related_model == exclude_model:
+                sub_queryset = sub_queryset.exclude(**exclude_kwargs)
+
+            final_value = []
+            for value in sub_queryset.distinct():
+                final_value.append(
+                    backend.get_search_field_transformation(
+                        search_field=related_field
+                    )(value) or ''
                 )
-                try:
-                    value = list(flatten_list(value=value))
-                    if value == [None]:
-                        value = None
-                    else:
-                        value = ' '.join(value)
-                except TypeError:
-                    """Value is not a list."""
-            except ResolverPipelineError:
-                """Not fatal."""
-            else:
-                result[field] = field_map[field].get(
-                    'transformation', SearchModel.function_return_same
-                )(value)
+
+            result[field_name] = ' '.join(final_value)
 
         return result