+++ mayan/apps/dynamic_search/tasks.py
@@ -1,18 +1,24 @@
 import logging
 
 from django.apps import apps
+from django.db.models.aggregates import Max, Min
 
 from mayan.apps.lock_manager.exceptions import LockError
 from mayan.celery import app
 
 from .classes import SearchBackend, SearchModel
-from .exceptions import DynamicSearchRetry
+from .exceptions import DynamicSearchException, DynamicSearchRetry
+from .literals import (
+    TASK_DEINDEX_INSTANCE_MAX_RETRIES, TASK_INDEX_INSTANCE_MAX_RETRIES
+)
+from .settings import setting_indexing_chunk_size
 
 logger = logging.getLogger(name=__name__)
 
 
 @app.task(
-    bind=True, ignore_result=True, max_retries=None, retry_backoff=True
+    bind=True, ignore_result=True,
+    max_retries=TASK_DEINDEX_INSTANCE_MAX_RETRIES, retry_backoff=True
 )
 def task_deindex_instance(self, app_label, model_name, object_id):
     logger.info('Executing')
@@ -31,21 +37,25 @@ def task_deindex_instance(self, app_label, model_name, object_id):
 @app.task(
     bind=True, ignore_result=True, max_retries=None, retry_backoff=True
 )
-def task_index_search_model(self, search_model_full_name):
+def task_index_search_model(self, search_model_full_name, range_string=None):
     search_model = SearchModel.get(name=search_model_full_name)
 
-    for instance in search_model.model._meta.default_manager.all():
-        task_index_instance.apply_async(
-            kwargs={
-                'app_label': instance._meta.app_label,
-                'model_name': instance._meta.model_name,
-                'object_id': instance.pk
-            }
-        )
+    kwargs = {
+        'range_string': range_string, 'search_model': search_model
+    }
+
+    try:
+        SearchBackend.get_instance().index_search_model(**kwargs)
+    except Exception as exception:
+        raise DynamicSearchException(
+            'Unexpected error calling `index_search_model` with '
+            'keyword arguments {}.'.format(kwargs)
+        ) from exception
 
 
 @app.task(
-    bind=True, ignore_result=True, max_retries=None, retry_backoff=True
+    bind=True, ignore_result=True,
+    max_retries=TASK_INDEX_INSTANCE_MAX_RETRIES, retry_backoff=True
 )
 def task_index_instance(
     self, app_label, model_name, object_id, exclude_app_label=None,
@@ -72,3 +82,24 @@ def task_index_instance(
         raise self.retry(exc=exception)
 
     logger.info('Finished')
+
+
+@app.task(ignore_result=True)
+def task_reindex_backend():
+    backend = SearchBackend.get_instance()
+    backend.reset()
+
+    for search_model in SearchModel.all():
+        queryset = search_model.model._meta.managers_map[search_model.manager_name].all()
+        if queryset:
+            queryset = queryset.aggregate(min_id=Min('id'), max_id=Max('id'))
+            step = setting_indexing_chunk_size.value - 1
+
+            for id_start in range(queryset['min_id'], queryset['max_id'] + 1, step):
+                range_string = '{}-{}'.format(id_start, id_start + step - 1)
+                task_index_search_model.apply_async(
+                    kwargs={
+                        'range_string': range_string,
+                        'search_model_full_name': search_model.get_full_name(),
+                    }
+                )