+++ mayan/apps/dynamic_search/tests/test_views.py
@@ -1,5 +1,3 @@
-from django.test import override_settings
-
 from mayan.apps.documents.models.document_models import DocumentSearchResult
 from mayan.apps.documents.permissions import permission_document_view
 from mayan.apps.documents.search import document_search
@@ -26,7 +24,6 @@ class AdvancedSearchViewTestCaseMixin(
         # Upload many instances of the same test document.
         for i in range(self.test_document_count):
             self._upload_test_document()
-            self._index_instance(instance=self.test_document)
             self.grant_access(
                 obj=self.test_document, permission=permission_document_view
             )
@@ -83,23 +80,28 @@ class AdvancedSearchViewTestCaseMixin(
             )
 
 
-@override_settings(SEARCH_BACKEND='mayan.apps.dynamic_search.backends.django.DjangoSearchBackend')
 class DjangoAdvancedSearchViewTestCase(
     AdvancedSearchViewTestCaseMixin, GenericViewTestCase
 ):
+    _test_search_backend_path = 'mayan.apps.dynamic_search.backends.django.DjangoSearchBackend'
     """Test against Django backend."""
 
 
-@override_settings(SEARCH_BACKEND='mayan.apps.dynamic_search.backends.whoosh.WhooshSearchBackend')
+class ElasticSearchAdvancedSearchViewTestCase(
+    AdvancedSearchViewTestCaseMixin, GenericViewTestCase
+):
+    _test_search_backend_path = 'mayan.apps.dynamic_search.backends.elasticsearch.ElasticSearchBackend'
+    """Test against ElasticSearch backend."""
+
+
 class WhooshAdvancedSearchViewTestCase(
     AdvancedSearchViewTestCaseMixin, GenericViewTestCase
 ):
+    _test_search_backend_path = 'mayan.apps.dynamic_search.backends.whoosh.WhooshSearchBackend'
     """Test against Whoosh backend."""
 
 
-class SearchViewTestCaseMixin(
-    DocumentTestMixin, SearchViewTestMixin,
-):
+class SearchViewTestCaseMixin(DocumentTestMixin, SearchViewTestMixin):
     def test_result_view_with_search_mode_in_data(self):
         self.grant_access(
             obj=self.test_document, permission=permission_document_view
@@ -116,24 +118,29 @@ class SearchViewTestCaseMixin(
         )
 
 
-@override_settings(SEARCH_BACKEND='mayan.apps.dynamic_search.backends.django.DjangoSearchBackend')
 class DjangoSearchViewTestCase(
     SearchViewTestCaseMixin, GenericViewTestCase
 ):
+    _test_search_backend_path = 'mayan.apps.dynamic_search.backends.django.DjangoSearchBackend'
     """Test against Django backend."""
 
 
-@override_settings(SEARCH_BACKEND='mayan.apps.dynamic_search.tests.backends.TestSearchBackend')
+class ElasticSearchSearchViewTestCase(
+    SearchViewTestCaseMixin, GenericViewTestCase
+):
+    _test_search_backend_path = 'mayan.apps.dynamic_search.backends.elasticsearch.ElasticSearchBackend'
+    """Test against ElasticSearch backend."""
+
+
 class WhooshSearchViewTestCase(
     SearchViewTestCaseMixin, GenericViewTestCase
 ):
+    _test_search_backend_path = 'mayan.apps.dynamic_search.backends.whoosh.WhooshSearchBackend'
     """Test against Whoosh backend."""
 
 
-@override_settings(SEARCH_BACKEND='mayan.apps.dynamic_search.tests.backends.TestSearchBackend')
-class SearchToolsViewTestCase(
-    DocumentTestMixin, SearchToolsViewTestMixin, SearchTestMixin,
-    GenericViewTestCase
+class SearchToolsViewTestCaseMixin(
+    DocumentTestMixin, SearchToolsViewTestMixin, SearchTestMixin
 ):
     def setUp(self):
         super().setUp()
@@ -143,7 +150,7 @@ class SearchToolsViewTestCase(
         )
 
     def test_search_backend_reindex_view_no_permission(self):
-        self.search_backend.clear_search_model_index(
+        self.search_backend.reset(
             search_model=self.document_search_model
         )
         self.grant_access(
@@ -161,7 +168,7 @@ class SearchToolsViewTestCase(
         self.assertEqual(queryset.count(), 0)
 
     def test_search_backend_reindex_view_with_permission(self):
-        self.search_backend.clear_search_model_index(
+        self.search_backend.reset(
             search_model=self.document_search_model
         )
         self.grant_access(
@@ -178,3 +185,24 @@ class SearchToolsViewTestCase(
             user=self._test_case_user
         )
         self.assertNotEqual(queryset.count(), 0)
+
+
+class DjangoSearchToolViewTestCase(
+    SearchToolsViewTestCaseMixin, GenericViewTestCase
+):
+    _test_search_backend_path = 'mayan.apps.dynamic_search.backends.django.DjangoSearchBackend'
+    """Test against Django backend."""
+
+
+class ElasticSearchToolViewTestCase(
+    SearchToolsViewTestCaseMixin, GenericViewTestCase
+):
+    _test_search_backend_path = 'mayan.apps.dynamic_search.backends.elasticsearch.ElasticSearchBackend'
+    """Test against ElasticSearch backend."""
+
+
+class WhooshSearchToolViewTestCase(
+    SearchToolsViewTestCaseMixin, GenericViewTestCase
+):
+    _test_search_backend_path = 'mayan.apps.dynamic_search.backends.whoosh.WhooshSearchBackend'
+    """Test against Whoosh backend."""